\name{Rbutchland-package}
\alias{Rbutchland-package}
\alias{Rbutchland}
\docType{package}
\title{
\packageTitle{Rbutchland}
}
\description{
\packageDescription{Rbutchland}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rbutchland}
\packageIndices{Rbutchland}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Rbutchland}

Maintainer: \packageMaintainer{Rbutchland}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
